refactor: MenuProducts 일급 컬렉션화하여 리팩토링
