+++ src/main/java/kitchenpos/menus/domain/Menu.java
@@ -29,14 +29,8 @@ public class Menu {
     @Column(name = "displayed", nullable = false)
     private boolean displayed;
 
-    @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
-    @JoinColumn(
-            name = "menu_id",
-            nullable = false,
-            columnDefinition = "binary(16)",
-            foreignKey = @ForeignKey(name = "fk_menu_product_to_menu")
-    )
-    private List<MenuProduct> menuProducts;
+    @Embedded
+    private MenuProducts menuProducts;
 
     @Transient
     private UUID menuGroupId;
@@ -85,11 +79,11 @@ public class Menu {
     }
 
     public List<MenuProduct> getMenuProducts() {
-        return menuProducts;
+        return this.menuProducts.getMenuProducts();
     }
 
     public void setMenuProducts(final List<MenuProduct> menuProducts) {
-        this.menuProducts = menuProducts;
+        this.menuProducts = new MenuProducts(menuProducts);
     }
 
     public UUID getMenuGroupId() {
@@ -101,30 +95,18 @@ public class Menu {
     }
 
     public void updateMenuProductPrice(UUID productId, BigDecimal productPrice) {
-        this.menuProducts.stream()
-                .filter(it -> it.getProductId().equals(productId))
-                .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException("상품을 찾을 수가 없습니다. productId : " + productId))
-                .updateProductPrice(productPrice);
-
-        var totalProductPrice = getTotalProductPrice();
-        if (this.menuPrice.isTotalProductPriceOver(totalProductPrice)) {
+        this.menuProducts.updateProductPrice(productId, productPrice);
+        if (this.menuPrice.isTotalProductPriceOver(this.menuProducts.getTotalProductPrice())) {
             this.displayed = false;
         }
     }
 
-    public BigDecimal getTotalProductPrice() {
-        return menuProducts.stream()
-                .map(MenuProduct::getProductPrice)
-                .reduce(BigDecimal.ZERO, BigDecimal::add);
-    }
-
     public void updateMenuPrice(BigDecimal price) {
-        this.menuPrice = menuPrice.updatePrice(price, this.getTotalProductPrice());
+        this.menuPrice = menuPrice.updatePrice(price, this.menuProducts.getTotalProductPrice());
     }
 
     public void display() {
-        if (this.menuPrice.isTotalProductPriceOver(this.getTotalProductPrice())) {
+        if (this.menuPrice.isTotalProductPriceOver(this.menuProducts.getTotalProductPrice())) {
             throw new IllegalStateException("상품 총 가격이 메뉴의 가격보다 높아 전시할 수 없습니다.");
         }
         this.displayed = true;
